refactor(iocuak): move convertBindingToBindingApi to an api folder
